% Generated by roxygen2 (4.0.2): do not edit by hand
\name{children}
\alias{children}
\alias{children.default}
\alias{children.wormsid}
\title{Retrieve immediate children taxa for a given taxon name or ID.}
\usage{
children(...)

\method{children}{default}(x, db = NULL, ...)

\method{children}{wormsid}(x, db = NULL, ...)
}
\arguments{
\item{...}{Further args passed on to \code{worms_children}.}

\item{x}{character; taxons to query.}

\item{db}{character; database to query. Only \code{worms} for now.}
}
\value{
A named list of data.frames with the children names of every supplied taxa.
You get an NA if there was no match in the database.
}
\description{
Retrieve immediate children taxa for a given taxon name or ID.
}
\examples{
\dontrun{
children("Aatolana", db = 'worms')
}
}

