% Generated by roxygen2 (4.0.2): do not edit by hand
\name{classification}
\alias{classification}
\alias{classification.default}
\alias{classification.wormsid}
\title{Retrieve the taxonomic hierarchy for a given taxon ID.}
\usage{
classification(...)

\method{classification}{default}(x, db = NULL, callopts = list(), ...)

\method{classification}{wormsid}(id, callopts = list(), ...)
}
\arguments{
\item{...}{Other arguments passed to \code{\link[taxize]{get_wormsid}}}

\item{x}{character; taxons to query.}

\item{db}{character; database to query. Only \code{worms} for now.}

\item{callopts}{Curl options passed on to \code{\link[httr]{GET}}}

\item{id}{character; identifiers, returned by \code{\link[taxize]{get_wormsid}}.}

\item{start}{The first record to return. If omitted, the results are returned
from the first record (start=0). This is useful if the total number of
  results is larger than the maximum number of results returned by a single
  Web service query (currently the maximum number of results returned by a
  single query is 500 for terse queries and 50 for full queries).}

\item{checklist}{character; The year of the checklist to query, if you want a specific
year's checklist instead of the lastest as default (numeric).}

\item{key}{Your API key; loads from .Rprofile.}
}
\value{
A named list of data.frames with the taxonomic classifcation of
   every supplied taxa.
}
\description{
Retrieve the taxonomic hierarchy for a given taxon ID.
}
\note{
If IDs are supplied directly (not from the \code{get_*} functions) you
   must specify the type of ID. There is a timeout of 1/3 seconds between
   querries to NCBI.
}
\examples{
\dontrun{
classification(x="Salvelinus fontinalis", db = 'worms')
classification(get_wormsid("Salvelinus fontinalis"))
}
}
\seealso{
\code{\link[taxize]{get_wormsid}}
}

