% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_ids_s}
\alias{get_ids_s}
\title{Retrieve taxonomic identifiers for a given taxon name.}
\usage{
get_ids_s(names, db = c("worms", "pesi"), ...)
}
\arguments{
\item{names}{character; Taxonomic name to query.}

\item{db}{character; database to query. One or  more of \code{worms} or \code{pesi}. By
default db is set to search all data sources.}

\item{...}{Other arguments passed to \code{\link[taxize]{get_wormsid}},
\code{\link[taxize]{get_pesiid}}.}
}
\value{
A vector of taxonomic identifiers, each retaining their respective S3
   classes so that each element can be passed on to another function (see e.g.'s).
}
\description{
This is a convenience function to get identifiers across all data sources. You can
use other \code{get_*} functions to get identifiers from specific sources if
you like.
}
\note{
There is a timeout of 1/3 seconds between queries to NCBI.
}
\examples{
\dontrun{
# Plug in taxon names directly
## By default you get ids for all data sources
get_ids_s(names="Chironomus riparius")

## Or you can specify which source you want via the db parameter
get_ids_s(names="Salvelinus fontinalis", db = 'worms')
}
}
\seealso{
\code{\link[taxize]{get_wormsid}}, \code{\link[taxize]{get_pesiid}}
}

