% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sci2comm}
\alias{sci2comm}
\alias{sci2comm.default}
\alias{sci2comm.wormsid}
\title{Get common names from scientific names.}
\usage{
sci2comm(...)

\method{sci2comm}{default}(scinames, db = "worms", simplify = TRUE, ...)

\method{sci2comm}{wormsid}(id, simplify = TRUE, ...)
}
\arguments{
\item{...}{Further arguments passed on to functions \code{\link[taxizesoap]{get_wormsid}}.}

\item{scinames}{character; One or more scientific names or partial names.}

\item{db}{character; Data source, only \emph{"worms"} for now}

\item{simplify}{(logical) If TRUE, simplify output to a vector of names. If FALSE,
return variable formats from different sources, usually a data.frame.}

\item{id}{character; identifiers, as returned by \code{\link[taxizesoap]{get_wormsid}}.}
}
\value{
List of character - vectors.
}
\description{
Get common names from scientific names.
}
\examples{
\dontrun{
sci2comm(scinames='Squatina squatina', db='worms')

# Passing id in, works for sources: itis and ncbi
sci2comm(get_wormsid('Squatina squatina'))

# Don't simplify returned
sci2comm(get_tsn('Helianthus annuus'), simplify=FALSE)
}
}
\author{
Scott Chamberlain (myrmecocystus@gmail.com)
}

