% Generated by roxygen2 (4.0.2): do not edit by hand
\name{synonyms}
\alias{synonyms}
\alias{synonyms.default}
\alias{synonyms.wormsid}
\title{Retrieve synonyms from various sources given input taxonomic names or identifiers.}
\usage{
synonyms(...)

\method{synonyms}{default}(x, db = NULL, ...)

\method{synonyms}{wormsid}(id, ...)
}
\arguments{
\item{...}{Other passed arguments.}

\item{x}{character; taxons to query.}

\item{db}{character; database to query. Only \code{worms} for now}

\item{id}{character; identifiers, returned by \code{\link[taxizesoap]{get_wormsid}}}
}
\value{
A named list of data.frames with the synonyms of every supplied taxa.
}
\description{
Retrieve synonyms from various sources given input taxonomic names or identifiers.
}
\note{
If IDs are supplied directly (not from the \code{get_*} functions) you
   must specify the type of ID.
}
\examples{
\dontrun{
# Plug in taxon names directly
synonyms("Salvelinus fontinalis", db='worms')

# Use methods for get_wormsid
synonyms(get_wormsid("Salvelinus fontinalis"))
}
}
\seealso{
\code{\link[taxizesoap]{get_wormsid}}
}

