% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{taxizesoap-package}
\alias{taxizesoap}
\alias{taxizesoap-package}
\title{Taxnomic search and phylogeny retrieval for SOAP data sources.}
\description{
Taxnomic search and phylogeny retrieval for SOAP data sources.
}
\details{
Currently supported APIs are:
\tabular{ll}{
World Register of Marine Species (WoRMS) \tab worms_ \cr
}
}
\keyword{package}

