% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worms_children.R
\name{worms_children}
\alias{worms_children}
\title{Children search of WoRMS data}
\usage{
worms_children(ids = NULL, offset = NULL, marine_only = 1, opts = NULL,
  iface = NULL, ...)
}
\arguments{
\item{ids}{(numeric) One or more WoRMS AphidID's for a taxon.}

\item{offset}{Starting record number, when retrieving next chunk of (50) records. Default=1.}

\item{marine_only}{(logical) Include results from marine taxa only. Default: TRUE.}

\item{opts}{(character) a named list of elements that are passed to the curlPerform function 
which actually invokes the SOAP method. These options control aspects of the HTTP request, 
including debugging information that is displayed on the console, 
e.g. .opts = list(verbose = TRUE)}

\item{iface}{Interface to WoRMS SOAP API methods. By default we use a previously created object.
If you want to create a new one, use \code{worms_gen_iface}, assign the output to an object, 
then pass it into any \code{worms_*} function. in the \code{iface} parameter.}

\item{...}{Further args passed on to \code{SSOAP::.SOAP}.}
}
\value{
A list of data.frame's, named by the input Aphia ids
}
\description{
Children search of WoRMS data
}
\examples{
\dontrun{
worms_children(ids=106135)
worms_children(ids=c(106135,159283))
out <- worms_children(ids=c(106135,159283))
lapply(out, head)

# to bind outputs together from multiple taxa
do.call(rbind, out)
## or with plyr::rbind.fill
do.call(rbind.fill, out)
}
}

