% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comm2sci_s}
\alias{comm2sci_s}
\title{Get scientific names from common names.}
\usage{
comm2sci_s(commnames, db = "worms", simplify = TRUE, ...)
}
\arguments{
\item{commnames}{One or more common names or partial names.}

\item{db}{Data source, only \emph{"worms"} for now}

\item{simplify}{(logical) If TRUE, simplify output to a vector of names. If FALSE,
return variable formats from different sources, usually a data.frame.}

\item{...}{Further arguments passed on to internal methods.}

\item{itisby}{Search for common names across entire names (search, default),
at beginning of names (begin), or at end of names (end).}
}
\value{
A vector of names.
}
\description{
Get scientific names from common names.
}
\examples{
\dontrun{
comm2sci_s(commnames=c('atlantic salmon','great white shark'), db='worms')
}
}
\author{
Scott Chamberlain (myrmecocystus@gmail.com)
}
\seealso{
\code{\link[taxize]{worms_records}}

\code{\link[taxize]{sci2comm}}
}

