% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxizesoap-package.R
\docType{package}
\name{taxizesoap-package}
\alias{taxizesoap-package}
\alias{taxizesoap}
\title{Taxnomic search and phylogeny retrieval for SOAP data sources.}
\description{
Taxnomic search and phylogeny retrieval for SOAP data sources.
}
\details{
Currently supported APIs are:
\tabular{ll}{
World Register of Marine Species (WoRMS) \tab worms_ \cr
}
}
\keyword{package}
